% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_chain.R
\name{calc_chain}
\alias{calc_chain}
\title{Calculate chain links and chain length}
\usage{
calc_chain(cs = NULL, chainring = NULL, sprocket = NULL, idler = 10)
}
\arguments{
\item{cs}{Length of the chain stay in mm.}

\item{chainring}{Number of teeth of the high gear chainring.}

\item{sprocket}{Number of teeth of the high gear sprocket.}

\item{idler}{Number of teeth of the idler roller, i.e. 10 or 11 (default:
10).}
}
\description{
Calculate the recommended number of chain links and the chain
  length in cm.
}
\examples{
calc_chain(cs = 400, chainring = 52, sprocket = 28)
}
\author{
Jannes Muenchow
}

