% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_fixie_tooths.R
\name{calc_fixie_tooths}
\alias{calc_fixie_tooths}
\title{Front and rear gear tooths (Fixie).}
\usage{
calc_fixie_tooths(ratio = 2.8, tol = 0)
}
\arguments{
\item{ratio}{The desired gear ratio given as a positive numeral.}

\item{tol}{Tolerance indicating the range around which the desired gear ratio
might deviate (default: 0).}
}
\value{
The function returns a dataframe with the ten best results in
  descending order.
}
\description{
\code{calc_fixie_tooths} calculates the optimal number of front
  and rear gear tooths for a given gear ratio/transmission.
}
\details{
The function assumes that the front sprocket has between 31 and 62
  tooths, and the rear sprocket between 14 and 21.
}
\examples{
calc_fixie_tooths(2.8, 0.1)
}
\author{
Jannes Muenchow
}
\keyword{Fixie}
\keyword{Single}
\keyword{bike,}
\keyword{speed}

