% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_skids.R
\name{calc_skids}
\alias{calc_skids}
\title{Skid patches}
\usage{
calc_skids(front, rear)
}
\arguments{
\item{front}{The number of chainring teeth.}

\item{rear}{The number of cog teeths.}
}
\value{
The function returns a \code{data.frame} with two columns:
\enumerate{
 \item{\strong{skid_1} The number of skid patches for single-legged
 skidders.}
 \item{\strong{skid_2} The number of skid patches for ambidextrous skidders
 (two-legged skidders).}
}
}
\description{
The function \code{calc_skids} calculates the number of skid
  patches for both single-legged and two-legged skidders for specified
  combinations of chainring- and cog-teeth.
}
\details{
The function was inspired by
  \url{http://www.bikecalc.com/skid_patch_math}. Basically, the function
  divides the cog by the the least common denominator of cog and chainring
  teeth. This corresponds to the number of skid patches for single-legged
  skidders. As an extra, the function also calculates the number of skid
  patches for ambidextrous skidders.
}
\examples{
calc_skids(54, 14)
# you can also specify vectors of length > 1
calc_skids(front = 48:54,
           rear = 12:18)
}

