% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_trail.R
\name{calc_trail}
\alias{calc_trail}
\title{Calculate trail}
\usage{
calc_trail(dia = NULL, rake = NULL, ha = NULL)
}
\arguments{
\item{dia}{Diameter of the wheel plus tire (in mm).}

\item{rake}{Rake also known as fork offest (in mm).}

\item{ha}{Head angle in degrees.}
}
\description{
Calculate the trail.
}
\examples{
calc_trail(dia = 700, rake = 58, ha = 73)
}
\author{
Dirk Haas, Jannes Muenchow
}

