% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_gcd}
\alias{find_gcd}
\title{Greatest common divisor}
\usage{
find_gcd(x = NULL, y = NULL)
}
\arguments{
\item{x}{an integer of length 1}

\item{y}{an integer of length 1}
}
\description{
Find the greatest common divisor of two integers.
}
\examples{
find_gcd(45, 15)
find_gcd(1289, 124)
}
\author{
Jannes Muenchow
}

