% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_multiplier}
\alias{find_multiplier}
\title{Find multiplier}
\usage{
find_multiplier(x)
}
\description{
Find lowest possible multiplier which yields an integer when
  multiplied with a numeric.
param x a numeric vector
importFrom stringr str_split
importFrom dplyr %>%
importFrom conf.design factorize
}
\examples{
find_multiplier(2.25)
find_multiplier(8.245)
find_multiplier(c(2.25, 8.245, 2.1))
}

