% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_wnr}
\alias{find_wnr}
\title{Find smallest equivalent whole number ratio}
\usage{
find_wnr(front, rear)
}
\arguments{
\item{front}{}

\item{rear}{}
}
\value{
a vector with the number of skid patches for single-legged skidders
importFrom conf.design factorize
}
\description{
Simplify the gear ratio to the smallest whole number ratio.
}

