% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_fixie_teeth.R
\name{calc_fixie_teeth}
\alias{calc_fixie_teeth}
\title{Optimal teeth combination (Fixie)}
\usage{
calc_fixie_teeth(ratio = 2.8, tol = 0.1, front = 31:59, rear = 14:21,
  nrow = 5)
}
\arguments{
\item{ratio}{The desired gear ratio given as a positive numeral.}

\item{tol}{Tolerance around which the specified gear ratio might deviate with
0.1 as default.}

\item{front}{Range of the number of chainring teeth given as an integer vector.}

\item{rear}{Range of the number of cog teeth given as an integer vector.}

\item{nrow}{An integer indicating the maximal number of rows which should be
returned by the function.}
}
\value{
The function returns a dataframe with following columns:
\enumerate{
\item{\strong{front} The optimal number of chainring teeth (see \code{Details}).}
\item{\strong{rear} The optimal number of cog teeth (see \code{Details}).}
\item{\strong{ratio} The actual gear ratio.}
\item{\strong{skid_1} The number of skid patches for single-legged skidders.
See \code{\link{calc_skids}} for more details.}
\item{\strong{skid_2} The number of skid patches for two-legged skidders.
See \code{\link{calc_skids}} for more details.}
}
}
\description{
\code{calc_fixie_teeth} calculates the optimal number of
  chainring (front) and cog (rear) teeth for a given gear ratio.
}
\details{
The function keeps only teeth combinations resulting in gear
  ratios within the specified tolerance range. Secondly, it sorts the results
  by \code{skid_2} in descending order.
}
\examples{
calc_fixie_teeth(2.8, 5)
}
\author{
Jannes Muenchow
}
\keyword{Fixie}
\keyword{Single}
\keyword{bike,}
\keyword{speed}

