% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_fixie_tooths.R
\name{calc_fixie_tooths}
\alias{calc_fixie_tooths}
\title{Front and rear gear tooths (Fixie).}
\usage{
calc_fixie_tooths(ratio = 2.8, top = 5)
}
\arguments{
\item{ratio}{The desired gear ratio given as a positive numeral.}

\item{top}{An integer indicating the maximal number of rows which should be
returned by the function.}
}
\value{
The function returns a dataframe with three columns:
\enumerate{
\item{\strong{front} The optimal number of front gear tooths.}
\item{\strong{rear} The optimal number of rear gear tooths.}
\item{\strong{ratio} The gear ratio.}
}
The function displays all possible combinations in descending order which
range +/- 0.1 around the desired gear ratio.
}
\description{
\code{calc_fixie_tooths} calculates the optimal number of front
and rear gear tooths for a given gear ratio/transmission.
}
\details{
The function assumes that the front sprocket has between 31 and 59
tooths, and the rear sprocket between 14 and 21.
}
\examples{
calc_fixie_tooths(2.8, 0.1)
}
\author{
Jannes Muenchow
}
\keyword{Fixie}
\keyword{Single}
\keyword{bike,}
\keyword{speed}

