% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_spoke_lengths.R
\name{calc_spoke_lengths}
\alias{calc_spoke_lengths}
\title{Calculation of spoke lengths}
\usage{
calc_spoke_lengths(n, erd, left_flange_d, right_flange_d, wl, wr, cross = 3,
  spoke_hole_d = 2.4, osb = 0)
}
\arguments{
\item{n}{Total number of spokes.}

\item{erd}{Effective rim diameter.}

\item{left_flange_d}{Left flange diameter.}

\item{right_flange_d}{Right flange diameter.}

\item{wl}{Width from center to left flange.}

\item{wr}{width from center to right flange}

\item{cross}{Number of crosses.}

\item{spoke_hole_d}{spoke hole diameter.}

\item{osb}{Offset spoke bed.}
}
\description{
Calculate the spoke lengths for a given erd, flange, etc.
}
\examples{
calc_spoke_lengths(n = 36, erd = 558, left_flange_d = 67,
                   right_flange_d = 67, wl = 44, wr = 31)
}

