% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_spoke_lengths.R
\name{calc_spoke_lengths}
\alias{calc_spoke_lengths}
\title{Calculation of spoke lengths}
\usage{
calc_spoke_lengths(n = 32, erd, left_flange_d, right_flange_d, wl, wr,
  cross = 3, spoke_hole_d = 2.4)
}
\arguments{
\item{n}{Total number of spokes per bike wheel (default: 32).}

\item{erd}{Effective rim diameter (mm), which is the diameter of the rim at
the nipple seats, sometimes also referred to as over spoke diameter.}

\item{left_flange_d}{Diameter of the left flange (mm).}

\item{right_flange_d}{Diameter of the right flange (mm).}

\item{wl}{Width from the hub center to the left flange (mm).}

\item{wr}{Width from the hub center to the right flange (mm).}

\item{cross}{Number of times each spoke crosses other spokes. The default is
3 (3-cross lacing pattern).}

\item{spoke_hole_d}{Diameter of the spoke hole (mm).}
}
\value{
The function returns a data.frame with one row and the columns
  \strong{left_length} and \strong{right_length} in mm.
}
\description{
Calculate the spoke lengths for a given effective rim diameter
  (ERD) in combination with the necessary hub dimensions.
}
\details{
Essentially, Euclidean geometry solves the problem of calculating
  the length of a spoke. ERD and hub dimensions provide the necessary lengths
  while crossing pattern and the number of spoke holes determine the angle.
  Please refer to
  \url{http://www.wheelpro.co.uk/support/spoke-length-proof.php} for an
  excellent mathematical proof, both in terms of explanation and depiction.
}
\note{
So far the function assumes that the nipple has a length of 12 mm.
}
\examples{
# Rigida Zac 2000 with 36 spoke holes
# Shimano Deore LX FH-M580 (rear hub)
calc_spoke_lengths(n = 36, erd = 537, left_flange_d = 45,
                   right_flange_d = 45, wl = 37.5, wr = 17.5)
}
\author{
Jannes Muenchow
}
\references{
Roger Musson (2013): A professional guide to wheelbuilding. 6th
  Edition. \url{http://www.wheelpro.co.uk/}.
}

