% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_multiplier}
\alias{find_multiplier}
\title{Find multiplier}
\usage{
find_multiplier(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Find lowest possible integer multiplier which yields an integer
  when multiplied by a decimal number.
}
\examples{
find_multiplier(2.25)
find_multiplier(8.245)
# x can also be a numeric vector of length > 1
find_multiplier(c(2.25, 8.245, 2.1))
}

