% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_wnr}
\alias{find_wnr}
\title{Find smallest equivalent whole number ratio}
\usage{
find_wnr(front, rear)
}
\arguments{
\item{front}{The number of chainring teeth.}

\item{rear}{The number of cog teeths.}
}
\value{
The function returns a vector specifying the number of skid patches
  for single-legged skidders.
}
\description{
Simplify the gear ratio to the smallest whole number ratio.
}
\details{
The function was inspired by
  \url{http://www.bikecalc.com/skid_patch_math}.
}
\examples{
find_wnr(54, 14)
# you can also specify a vector
find_wnr(front = 48:54,
         rear = 12:18)
}

